First Try At MVC Chapter
Added Content to the MVC chapter that describes what MVC is and why it
is important.  Left the implimentation details to the Model, View and
Controller Chapters.

Need to add Links to spesific chapters.  It would be nice to be able to
add markup to link internally in the document withou having to depend on
the URI scheme.

Should also look for good Internet See Also refernces.  I Suggest
linking to the Rails Envy Public Service Announcments.