Generate Application Draft Completed
Filled in descriptions of the other three merb-gen application forms using the
pattern in very-flat.  Core team should check for accuracy.  May want to add
descriptions of when to use which form.

Reorded from biggest to smallest antisipation that readers will be most
interested in generating the full stack on their initial read-through.