First Half of 2.5 Controllers (en)
About 1600 words on Controllers including an outline for the rest of the
text.  Text is WordWrapped at 80 chars.

Includes an outline for the rest of the chapter.

"In this chapter we will look at how organize controllers using two of
the common controller styles, discuss how to write the methods that are
called be the framework on a controller, and how to extend the
functionality of a controller using different formats, filter,
inheritance and private methods."