2.8 Request Path
First draft of this chapter.  People with Jave experence need to read it
and add to/confirm the information in it.There are some links that need
to be filled in as appropriate.

Formatted as per Guidelines.

"All web applications begin with an HTTP request by a client. They all
end with a response from our server. In this chapter, we are
going to follow the program flow from the client request through to the
response."